requirements.txt
pandas>=2.0.0
numpy>=1.24.0
kaggle>=1.5.0
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.13.0
nltk>=3.8.1
scikit-learn>=1.2.0
wordcloud>=1.9.0
snowflake-connector-python>=3.0.0
openai>=1.0.0
python-dotenv>=1.0.0
langchain>=0.0.300
streamlit>=1.25.0
chroma-db>=0.4.0
